restructure tests ahead of adding archive specific packages
